% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vgvi.R
\name{calculate_weights}
\alias{calculate_weights}
\title{Calculate Weights Based on Distance and Decay Mode}
\usage{
calculate_weights(distance_raster, mode, m, b)
}
\arguments{
\item{distance_raster}{A `SpatRaster` object representing distances from a specific point or points. Each cell's value represents the distance to the point(s) of interest.}

\item{mode}{A character string specifying the type of decay function to apply. Supported modes are "logistic" and "exponential". If any other value is provided, the function returns a weight of 1 for all distances, implying no decay.}

\item{m}{A numeric value used as a parameter in the decay functions. For the logistic mode, `m` adjusts the midpoint of the sigmoid curve where the weight is 0.5. For the exponential mode, `m` influences the rate at which weights decrease with distance.}

\item{b}{A numeric value used as a parameter in the decay functions to control the steepness (logistic) or the scale (exponential) of the decay.}
}
\value{
A `SpatRaster` object where each cell's value represents the calculated weight based on the specified decay mode and the distance to the point(s) of interest.
}
\description{
This internal function calculates weights for each cell in a distance raster based on the specified decay mode. The function supports logistic and exponential decay modes, applying a decay function to the normalized distance to compute weights. This function is typically used internally within spatial analysis functions to model the diminishing influence of features (e.g., greenspaces) with increasing distance from a point of interest.
}
\keyword{internal}
