% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gavi.R
\name{gavi}
\alias{gavi}
\title{Greenspace Availability Index (GAVI)}
\usage{
gavi(x, lac, na.rm = TRUE, cores = 1, progress = FALSE)
}
\arguments{
\item{x}{A `SpatRaster` object.}

\item{lac}{A data frame containing lacunarity information with columns (see (\code{\link[CGEI]{lacunarity}}).}

\item{na.rm}{A logical indicating whether NA values should be removed.}

\item{cores}{The number of cores to use for parallel processing. Default is 1.}

\item{progress}{logical; Show progress bar?}
}
\value{
A `SpatRaster` object representing the GAVI.
}
\description{
This function computes the Greenspace Availability Index (GAVI) from a lacunarity dataset.
}
\details{
For every layer, the focal means of all box sizes in \code{lac} are weighted by their lacunarity
and averaged. These values are classified into 9 classes using Jenks natural breaks ("fisher"
style for rasters with more than 5,000 cells, "jenks" otherwise). The GAVI is the mean of the
classes of all layers, again classified into 9 classes if \code{x} has more than one layer.
Cells that are NA in a layer are NA in that layer's classes (and in the GAVI).
The natural breaks are computed from all non-NA cells or, if there are more than 50,000 of them,
from a random sample of 50,000 cells; use \code{set.seed()} for reproducible results.
}
\examples{
library(CGEI)
library(terra)

mat_sample <- matrix(data = c(
  1,1,0,1,1,1,0,1,0,1,1,0,
  0,0,0,0,0,1,0,0,0,1,1,1,
  0,1,0,1,1,1,1,1,0,1,1,0,
  1,0,1,1,1,0,0,0,0,0,0,0,
  1,1,0,1,0,1,0,0,1,1,0,0,
  0,1,0,1,1,0,0,1,0,0,1,0,
  0,0,0,0,0,1,1,1,1,1,1,1,
  0,1,1,0,0,0,1,1,1,1,0,0,
  0,1,1,1,0,1,1,0,1,0,0,1,
  0,1,0,0,0,0,0,0,0,1,1,1,
  0,1,0,1,1,1,0,1,1,0,1,0,
  0,1,0,0,0,1,0,1,1,1,0,1
), nrow = 12, ncol = 12, byrow = TRUE)

x <- rast(mat_sample)
x2 <- rast(mat_sample*runif(144, 1, 2))

x <- c(x, x2)
lac <- lacunarity(x)
gavi(x, lac)

}
