% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewshed.R
\name{viewshed_list}
\alias{viewshed_list}
\title{Viewshed}
\usage{
viewshed_list(
  observer,
  dsm_rast,
  dtm_rast,
  max_distance = 200,
  observer_height = 1.7,
  spacing = NULL,
  cores = 1L,
  progress = FALSE
)
}
\arguments{
\item{observer}{An `sf` object containing the observer locations. Each observer location should be a point geometry with a defined coordinate reference system (CRS).}

\item{dsm_rast}{A `SpatRaster` object representing the Digital Surface Model (DSM) of the area, indicating the elevation of surface objects including vegetation, buildings, and other features.}

\item{dtm_rast}{A `SpatRaster` object representing the Digital Terrain Model (DTM) of the area, indicating the elevation of the ground surface without any objects.}

\item{max_distance}{The maximum distance (in meters) from the observer location within which the greenspace visibility is calculated. Default is 200 meters.}

\item{observer_height}{The height (in meters) of the observer above the ground level. Default is 1.7 meters.}

\item{spacing}{optional; numeric > 0; Only if \code{observer} is a LINESTRING or POLYGON. Points on the line will be generated. The \code{spacing} parameter sets the distance in between the points on the line/grid. Defaults to the resolution of the \code{dsm_rast}.}

\item{cores}{The number of cores to use for parallel processing. Default is 1.}

\item{progress}{logical; Show progress bar and computation time?}
}
\value{
List of `SpatRaster` objects with the viewshed results.
}
\description{
Computes a binary viewshed of a point on a Digital Surface Model (DSM). The observer height is based on the heigth of a Digital Terrain Model (DTM).
}
\details{
observer needs to be a geometry of type POINT, LINESTRING, MULTILINESTRING, POLYGON or MULTIPOLYGON. If observer is a LINESTRING or MULTILINESTRING, 
points will be generated along the line(s) every "spacing" meters. If observer is a POLYGON or MULTIPOLYGON, a grid with resolution = "spacing" 
will be generated, and a viewshed will be computed for every point.
The CRS (\code{\link[sf]{st_crs}}) needs to have a metric unit!
}
\examples{
\dontrun{
library(sf)
library(terra)

# Simulate observer locations as sf points
observer <- st_as_sf(data.frame(lon = c(5, 5.1), lat = c(52, 52.1)),
                      coords = c("lon", "lat"), crs = 4326)

# Transform to UTM zone 31N for metric units
observer <- st_transform(observer, 32631)

# Create synthetic rasters for DSM, DTM, and greenspace around observers
# This is a simplified example assuming flat terrain and random greenspaces
bbox_observer <- st_bbox(observer)
x_range <- bbox_observer[c("xmin", "xmax")]
y_range <- bbox_observer[c("ymin", "ymax")]

# Create a raster with 1km x 1km around the observer
dsm_rast <- rast(res=100,
                 xmin=min(x_range)-1000, xmax=max(x_range) + 1000, 
                 ymin=min(y_range)-1000, ymax=max(y_range) + 1000, 
                 crs=crs(observer))
dsm_rast[] <- runif(ncell(dsm_rast), 0, 1.5) # Assign random heights

dtm_rast <- rast(dsm_rast, vals=0) # Flat terrain

# Calculate viewshed
viewshed_results <- viewshed(observer, dsm_rast, dtm_rast)
}

}
