% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gavi.R
\name{gavi}
\alias{gavi}
\title{Greenspace Availability Index (GAVI)}
\usage{
gavi(x, lac, na.rm = TRUE, cores = 1, progress = FALSE)
}
\arguments{
\item{x}{A `SpatRaster` object.}

\item{lac}{A data frame containing lacunarity information with columns (see (\code{\link[CGEI]{lacunarity}}).}

\item{na.rm}{A logical indicating whether NA values should be removed.}

\item{cores}{The number of cores to use for parallel processing. Default is 1.}

\item{progress}{logical; Show progress bar?}
}
\value{
A `SpatRaster` object representing the GAVI.
}
\description{
This function computes the Greenspace Availability Index (GAVI) from a lacunarity dataset.
}
\examples{
library(CGEI)
library(terra)

mat_sample <- matrix(data = c(
  1,1,0,1,1,1,0,1,0,1,1,0,
  0,0,0,0,0,1,0,0,0,1,1,1,
  0,1,0,1,1,1,1,1,0,1,1,0,
  1,0,1,1,1,0,0,0,0,0,0,0,
  1,1,0,1,0,1,0,0,1,1,0,0,
  0,1,0,1,1,0,0,1,0,0,1,0,
  0,0,0,0,0,1,1,1,1,1,1,1,
  0,1,1,0,0,0,1,1,1,1,0,0,
  0,1,1,1,0,1,1,0,1,0,0,1,
  0,1,0,0,0,0,0,0,0,1,1,1,
  0,1,0,1,1,1,0,1,1,0,1,0,
  0,1,0,0,0,1,0,1,1,1,0,1
), nrow = 12, ncol = 12, byrow = TRUE)

x <- rast(mat_sample)
x2 <- rast(mat_sample*runif(144, 1, 2))

x <- c(x, x2)
lac <- lacunarity(x)
gavi(x, lac)

}
